+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 #### Fixed
 * **\[processor\]** Updated the `isDispose()` method to avoid invoking `reportObserved()` on the `"disposable"`
   observable property if the component is disposed or being disposed.
+* **\[processor\]** Suppressed `unchecked` warnings due to casts in the `@Memoize` methods with type parameters
+  generated in the enhanced component subclass.
 
 ### [v0.55](https://github.com/arez/arez/tree/v0.55) (2018-02-20)
 [Full Changelog](https://github.com/arez/arez/compare/v0.54...v0.55)